pyproject.toml
[project]
name = "this-is-fine"
version = "0.1.0"
description = "fire risk monitor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "dash>=3.2.0",
    "duckdb>=1.4.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "geopandas>=1.1.1",
    "httpx>=0.28.1",
    "keplergl>=0.3.7",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "prefect>=3.4.23",
    "pyproj>=3.7.2",
    "python-dotenv>=1.1.1",
    "shapely>=2.1.2",
]

[project.scripts]
wildfire = "src.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "basedpyright>=1.31.7",
    "jupyter>=1.1.1",
    "pandas-stubs>=2.3.2.250926",
    "pytest>=8.4.2",
]
